mcp-demo-backend/requirements.txt
fastapi
uvicorn
mcp
anthropic
python-dotenv
pydantic
